+++ NEWS.md
@@ -1,7 +1,9 @@
 # RmdConcord 0.2.1
 
-* Support `markdown::html_format` output via new
-`html_formatC()` function.
+* Added support for `markdown::html_format()` output via new
+`html_formatC()` function.  
+* Support for `markdown::latex_format()` is not planned.
+See `?html_formatC()` for some discussion of the issues.
 
 # RmdConcord 0.2.0